feat: add stopMin/stopMax/stopUnit props for custom gradient stop range
The gradient STOP input now displays values remapped onto [stopMin, stopMax]
with an optional unit rendered after the value, instead of always showing the
raw 0-100% position. Defaults (0, 100, no unit) preserve existing behavior.

+++ src/utils/gradientParser.ts
@@ -164,7 +164,7 @@ export const gradientParser = (input = '') => {
       error('Expected color definition')
     }
 
-    color.left = parseInt(matchDistance()?.value)
+    color.left = parseFloat(matchDistance()?.value)
     return color
   }
